compile

+++ tests/my_test-1.cpp
@@ -5,7 +5,7 @@
 using namespace std;
 
 int main() {
-    WeightedGraph<int> g;
+    WeightedGraph<double> g;
 
     assert(g.AddVertex(1) == true);
     assert(g.AddVertex(2) == true);
